# 8290824: Use InputStream.readAllBytes() instead of explicit loops