--- a/src/java.base/share/classes/sun/security/provider/certpath/X509CertPath.java
+++ b/src/java.base/share/classes/sun/security/provider/certpath/X509CertPath.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2000, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2000, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -26,6 +26,7 @@
 package sun.security.provider.certpath;
 
 import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.security.cert.CertificateEncodingException;
@@ -125,7 +126,7 @@ public class X509CertPath extends CertPath {
         // and the methods in the Sun JDK 1.4 implementation of ArrayList that
         // allow read-only access are thread-safe.
         this.certs = Collections.unmodifiableList(
-                new ArrayList<>((List<X509Certificate>)certs));
+                new ArrayList<X509Certificate>((List<X509Certificate>)certs));
     }
 
     /**
@@ -184,14 +185,14 @@ public class X509CertPath extends CertPath {
         }
 
         try {
-            DerInputStream dis = new DerInputStream(is.readAllBytes());
+            DerInputStream dis = new DerInputStream(readAllBytes(is));
             DerValue[] seq = dis.getSequence(3);
             if (seq.length == 0) {
-                return Collections.emptyList();
+                return Collections.<X509Certificate>emptyList();
             }
 
             certFac = CertificateFactory.getInstance("X.509");
-            certList = new ArrayList<>(seq.length);
+            certList = new ArrayList<X509Certificate>(seq.length);
 
             // append certs in reverse order (target to trust anchor)
             for (int i = seq.length-1; i >= 0; i--) {
@@ -227,7 +228,7 @@ public class X509CertPath extends CertPath {
             if (is.markSupported() == false) {
                 // Copy the entire input stream into an InputStream that does
                 // support mark
-                is = new ByteArrayInputStream(is.readAllBytes());
+                is = new ByteArrayInputStream(readAllBytes(is));
             }
             PKCS7 pkcs7 = new PKCS7(is);
 
@@ -237,7 +238,7 @@ public class X509CertPath extends CertPath {
                 certList = Arrays.asList(certArray);
             } else {
                 // no certs provided
-                certList = new ArrayList<>(0);
+                certList = new ArrayList<X509Certificate>(0);
             }
         } catch (IOException ioe) {
             throw new CertificateException("IOException parsing PKCS7 data: " +
@@ -248,6 +249,22 @@ public class X509CertPath extends CertPath {
         // and the methods in the Sun JDK 1.4 implementation of ArrayList that
         // allow read-only access are thread-safe.
         return Collections.unmodifiableList(certList);
+    }
+
+    /*
+     * Reads the entire contents of an InputStream into a byte array.
+     *
+     * @param is the InputStream to read from
+     * @return the bytes read from the InputStream
+     */
+    private static byte[] readAllBytes(InputStream is) throws IOException {
+        byte[] buffer = new byte[8192];
+        ByteArrayOutputStream baos = new ByteArrayOutputStream(2048);
+        int n;
+        while ((n = is.read(buffer)) != -1) {
+            baos.write(buffer, 0, n);
+        }
+        return baos.toByteArray();
     }
 
     /**