We can use `InputStream.readAllBytes()` in `ModuleHashes` and `X509CertPath`.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [ ] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8290824](https://bugs.openjdk.org/browse/JDK-8290824): Use InputStream.readAllBytes() instead of explicit loops


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/9596/head:pull/9596` \
`$ git checkout pull/9596`

Update a local copy of the PR: \
`$ git checkout pull/9596` \
`$ git pull https://git.openjdk.org/jdk pull/9596/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 9596`

View PR using the GUI difftool: \
`$ git pr show -t 9596`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/9596.diff">https://git.openjdk.org/jdk/pull/9596.diff</a>

</details>