@mbreinhold 2022-07-21
Why do you say “instead of surrogate code”? I don’t see any Unicode surrogate processing here, but maybe you mean something else. Would “instead of explicit loops” be a better summary?