@stsypanov 2022-07-22
> Please revise the summary to say “... instead of explicit loops”.

Done.

> Please measure this (using, e.g., [JMH](https://github.com/openjdk/jmh)) before making such a change.

I'm aware of JMH, I was asking mostly about public API I should call in a benchmark to execute the changed code.