@stsypanov 2022-07-22
Could one advise how to execute smth like `jmod hash --module-path ~/openjdk-18.0.2/Contents/Home/jmods --hash-modules .*`? This command doesn't work for me.