@turbanoff 2022-07-23
Ow. I've created similar PR for `X509CertPath` recently #9263